for/taking-a-look-inside/src/requirements.txt
Pillow
pycrypto
pyscreenshot
